+++ R/join.R
@@ -25,7 +25,8 @@
 #' @export
 #'
 #' @examples
-#' if (requireNamespace("dplyr", quietly = TRUE)) {
+#' if (requireNamespace("dplyr", quietly = TRUE) &&
+#'     nrow(available_jpmap_data()) > 0) {
 #'   data("jp_prefecture_gdp")
 #'
 #'   gdp <- jp_prefecture_gdp |>