+++ R/leaflet.R
@@ -40,7 +40,8 @@
 #'
 #' @examples
 #' if (requireNamespace("leaflet", quietly = TRUE) &&
-#'     requireNamespace("dplyr", quietly = TRUE)) {
+#'     requireNamespace("dplyr", quietly = TRUE) &&
+#'     nrow(available_jpmap_data()) > 0) {
 #'   data("jp_prefecture_gdp")
 #'
 #'   gdp <- jp_prefecture_gdp |>